renamed component files to snake case
