+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import './globals.css';
-import NavBar from '@/components/NavBar/NavBar';
+import NavBar from '@/app/components/nav_bar/nav_bar';
 import Providers from './providers';
 
 const inter = Inter({ subsets: ['latin'] });
