+++ app/providers.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { SessionProvider } from 'next-auth/react';
-import { ThemeProvider } from '@/components/ThemeProvider';
+import { ThemeProvider } from '@/app/components/theme_provider';
 
 export default function Providers({ children }: { children: React.ReactNode }) {
   return (
